BREAD Working Paper 635 - The Social Consequences of Traditional Religion in Contemporary Africa
In sub-Saharan Africa, despite the adoption of Christianity, traditional religious beliefs remain widely held. We examine the social consequences of holding traditional religious beliefs among urban and rural populations in central Africa. Using a variety of lab-in-the-field experiments that randomize partner characteristics, we test whether individuals who believe in traditional religion are treated or viewed differently by others. We find that participants act less prosocially towards partners known to hold traditional religious beliefs. We find that this behavior is supported by norms and by negative perceptions and stereotypes of traditional believers. The effects are economically important, ubiquitous, and are amplified by economic development. Individual-level data from across the African continent reveal patterns consistent with our experimental findings. Individuals who believe in witchcraft have lower incomes, and the effect is stronger in countries that are more developed. Our final analysis speaks to the origins of these effects. Within our experimental sample, the negative effects are stronger in rural villages with more historical missionary activity, and across the continent, the negative relationship between belief in witchcraft and income is stronger in regions with more colonial missionary presence. Both findings are consistent with descriptive accounts of Christianity leading to the stigmatization of African traditional religion.