WP 627: Impact of forced labor migration on fertility west africa
s the persistently high fertility in West Africa today rooted in the decades of forced labor migration under colonial rule? We study the case of Burkina Faso, considered the largest labor reservoir in West Africa by the French colonial authorities. Hundreds of thousands of young men were forcibly recruited and sent to work in neighboring colonies for multiple years. The practice started in the late 1910s and lasted until the late 1940s, when forced labor was replaced with voluntary wage employment. We digitize historical maps, combine data from multiple surveys, and exploit the historical, temporary partition of colonial Burkina Faso (and, more specifically, the historical land of the Mossi ethnic group) into three zones with different needs for labor to implement a spatial regression discontinuity design analysis. We find that, on the side where Mossi villages were more exposed to forced labor historically, there is more temporary male migration to Cote d’Ivoire up to today, and lower realized and desired fertility today. We show evidence suggesting that the inherited pattern of low-skill circular migration for adult men reduced the reliance on subsistence farming and the accompanying need for child labor. We can rule out women’s empowerment or improvements in human and physical capital as pathways for the fertility decline. These findings contribute to the debate on the origins of family institutions and preferences, often mentioned to explain West Africa’s exceptional fertility trends, showing that fertility choices respond to changes in modes of production.