WP 636 - Lab Experiments in developing country contexts
Abstract: Lab-in-the-field experiments, in which lab experiments are conducted in more naturalistic settings, are increasingly being implemented in developing country contexts. In this chapter, we outline the conceptual and logistical challenges typically associated with lab-in-the-field experiments in non-Western settings. We describe the importance of worldviews and how researcher preconceptions may inadvertently shape the types of research questions that are asked. We emphasize the importance of increasing diversity of subject pools and researchers. We also suggest a set of best practices when implementing lab-in-the-field experiments in developing countries...