Add files via upload
Abstract: We survey the recent literature studying the effects of globalization on inequality in Latin America. Our focus is on research emerging from the late 2000s onward, with an emphasis on empirical work considering new mechanisms, studying new dimensions of inequality, and developing new methodologies to capture the many facets of globalization’s relationship to inequality. After summarizing both design-based and quantitative work in this area, we propose directions for future work. Our overarching recommendation is that researchers develop unifying frameworks to help synthesize the results of individual studies that focus on distinct aspects of globalization’s relationship to inequality.