WP 621 - On the Importance of African Traditional Religion for Economic Behavior
Within the field of economics, despite being widespread, African traditional religions tend to be perceived as unimportant and ignored when studying economic decision-making. This study tests whether this presumption is correct. Using daily data on business decisions and performance of beer sellers in the eastern Democratic Republic of the Congo, we study the importance of traditional religious beliefs for economic behavior and outcomes. Beer sellers perceive the risk of theft in their shops to be higher than it actually is, causing them to hold lower inventories, more frequent stock-outs, and reduced profits. We facilitate randomly-timed access to commonly-used, but typically prohibitively expensive rituals, which reduce the perceived risk of theft. We find that the rituals partially correct the beliefs about the risk of theft for sellers who report believing in the ritual’s efficacy. These sellers purchase more inventory, experience fewer stock-outs, and have larger sales, revenues, and profits. To distinguish the belief in the efficacy of the ritual from other incidental effects of participation, we analyze these outcomes for sellers who do not believe in the ritual. For these individuals, we find none of the observed effects. The findings provide evidence of the importance of African traditional religions, demonstrating that they can influence behavior and outcomes that are important for economic development.