WP 614 - Scaling Agricultural Policy Interventions
Policies aimed at raising agricultural productivity have been a centerpiece in the fight against global poverty. Their impacts are often measured using field or quasi-experiments that provide strong causal identification, but may be too small-scale to capture the general equilibrium (GE) effects that emerge once the policy is scaled up to a broader segment of the population. We pro- pose a new approach for quantifying large-scale GE policy counterfactuals that can both complement and be informed by evidence from field and quasi-experiments in agricultural settings. We develop a quantitative model of farm production, consumption and trading that captures important features of this setting, and propose a new solution method that relies on rich but widely available microdata. We showcase our approach in the context of a subsidy for modern inputs in Uganda, using administrative data for model calibration and variation from field and quasi-experiments for parameter estimation. We find that both the average and distributional impacts of the subsidy differ meaningfully when comparing a local intervention to one at scale, even for the same sample of farmers, and quantify the underlying mechanisms. We further document new insights on how the sign and extent of GE forces differ as a function of saturation rates at different geographical scales, and on the importance of capturing a granular economic geography for counterfactual analysis. Finally, we discuss practical considerations for combining our toolkit with evidence from field and quasi-experiments.