WP 613 - Imperfect Competition and Sanitation: Evidence from Randomized Auctions in Senegal
We document the impact of imperfect competition in the sanitation market in Dakar Senegal, both in the traditional market and in an experimental, just-in-time auction system that we designed in collaboration with the government. The auction platform randomly assigns each job to a set of bidders and to an auction format, either sealed bid or revisable bid (much like a sequential auction with open followed by closed bidding). We identify different collusive strategies in the two auction formats, and show that there are many participants who choose those different strategies when bidding in auctions of different formats. We discuss and rule out explanations other than collusion for the sub-optimal strategies. Collusion leads to higher prices and lower take-up rates for improved sanitation services, with implications for health.