WP 608- The Rise and Fall of SES Gradients in Heights around the World
We use data from a large sample of low and middle income countries to study the association (or “gradient”) between child height and maternal education. While the strong positive association between child health and measures of parental socio-economic status (SES) is well established, we uncover novel results regarding the evolution of this gradient as children age. The association is small at birth, rises throughout childhood and declines in adolescence as girls and boys approach puberty. This pattern is consistent with a degree of catch up in height among children of low SES families, in partial contrast to the argument that height deficits cannot be overcome after the early years of life. This catch up is partly explained by the association between SES and the timing of puberty and therefore of the adolescent growth spurt. By contrast, we do not find evidence in support of the role of behavioral responses in driving the inverted U-shape of the gradient. JEL: I14, I15, O15 Key words: Height, Socio-Economic Status, Maternal Schooling