WP 607- Estimating the Economic Value of Zoning Reform
We develop a framework to estimate the economic value of a recent zoning reform in Sao Paulo. Using a block-level regression discontinuity design, we find that developers request more permits in blocks with higher allowable densities. We incorporate these micro-estimates into an equilibrium model of housing supply and demand, finding that the reform produces a 1.9% increase in housing stock and a 0.5% reduction in prices, with substantial heterogeneity across neighborhoods. Welfare gains increase 4-fold once accounting for changes in built environment, and gains are larger for high-income and high-education families. However, homeowner house price losses overshadow all consumer gains.