WP 606- On the Dynamics of Human Behavior: The Past, Present, and Future of Culture, Conflict, and Cooperation
I provide a theoretically-guided discussion of the dynamics of human behavior, focusing on the importance of culture (socially-learned information) and tradition (transmission of culture across generations). Decision-making that relies on tradition can be an effective strategy and arises in equilibrium. While dynamically optimal, it generates static ‘mismatch.’ When the world changes, since traits evolve slowly, they may not be beneficial in their new environment. I discuss how mismatch helps explain the world around us, presents special challenges and opportunities for policy, and provides important lessons for our future as a human species.