WP 604- Democratization, Elite Capture and Economic Development
We show using a theoretical framework that embeds a voting model in a general-equilibrium model of a rural economy with two interest groups defined by land ownership that the effects of democratization – a shift from control of public resources by the landed elite to a democratic regime with universal suffrage – on the portfolio of public goods is heterogeneous, depending the population landless. In accord with the model and empirical findings from micro data on the differing material interests of the two land classes, we find, based on 30-year panel data describing the democratization of Indian villages, that democratization in villages with a larger landless population share shifted resources away from public irrigation, secondary schools, and electrification and towards programs that increase employment. When the landed farmers have a large population share, public resources were shifted towards irrigation, secondary schools and electrification and away from employment programs.