WP 601- Firm-Level Upgrading in Developing Countries
In principle, firms in developing countries benefit from the fact that advanced technologies and products have already been developed in industrialized countries and can simply be adopted, a process often referred to as industrial upgrading. But for many firms this advantage remains elusive. What is getting in the way? This paper reviews recent firm-level empirical research on the determinants of upgrading in developing countries. The first part focuses on how to define and measure various dimensions of upgrading — learning, quality upgrading, technology adoption, and product innovation. The second part takes stock of recent micro-empirical evidence on the drivers of upgrading, classifying them as output-side drivers, input-side drivers, or drivers of know-how. The review concludes with some thoughts about promising directions for research in the area.