WP 600- Why do People Stay Poor?
There are two broad views as to why people stay poor. One emphasizes differences in fundamentals, such as ability, talent or motivation. The other, the poverty traps view, differences in opportunities which stem from access to wealth. To test between these two views, we exploit a large-scale, randomized asset transfer and an 11-year panel on 6000 households who begin in extreme poverty. The setting is rural Bangladesh and the asset is cows. The data supports the poverty traps view – we identify a threshold level of initial assets above which households accumulate assets, take on better occupations (from casual labor in agriculture or domestic services to running small livestock businesses) and grow out of poverty. The reverse happens for those below the threshold. Structural estimation of an occupational choice model reveals that almost all beneficiaries are misallocated in the work they do at baseline and that the gains arising from eliminating misallocation would far exceed the program costs. Our findings imply that large transfers which create better jobs for the poor are an effective means of getting people out of poverty traps and reducing global poverty.