WP 598- Privatization of Public Goods: Evidence from the Sanitation Sector in Senegal
Privatization of a public good (the management of sewage treatment centers in Dakar, Senegal) leads to an increase in the productivity of downstream sewage dumping companies and a decrease in downstream prices of the services they provide to households. We use the universe of legal dumps of sanitation waste from May 2009 to November 2018 to show that legal dumping increased substantially following privatization – on average an increase of 61.7%, or an increase of about 14,000 cubic meters each month. This is due to increased productivity of all trucks, not just those associated with the company managing the treatment centers. Household-level survey data shows that prices of legal sanitary dumping decreased by ten percent following privatization, and DHS data show that diarrhea rates among children under five decreased in Dakar relative to secondary cities in Dakar following privatization with no similar effect on respiratory illness as a placebo.