WP 597- Persuading Voters to Punish Corrupt Vote Buyers: Experimental Evidence from a Large-Scale Radio Campaign in India
During the 2014 Indian general elections, we carried out a large-scale experiment randomizing a radio campaign highlighting the disadvantages of voting for corrupt vote-buying politicians. Official electoral data shows that the radio campaign significantly decreased the vote share of parties that engaged in the most vote buying (as reported by journalists). Voter survey data shows that the campaign increased the salience of government corruption as an election issue and decreased voting for parties that offered gifts. From a policy perspective, we show that radio campaigns are a cost-effective method to influence voter behavior. From a measurement perspective, we show that journalist interviews can impartially identify vote-buying politicians.