WP 592- Women Left Behind: Gender Disparities in Utilization of Government Health Insurance in India
Using administrative data on over 4 million hospital visits, we document striking gender disparities within a government health insurance program that entitles 46 million poor individuals to free hospital care in Rajasthan, India. Females account for only 33% of hospital visits among children and 43% among the elderly. These shares are lower for more expensive types of care, and far lower than sex differences in illness prevalence can explain. Almost two-thirds of non-childbirth spending is on males. We combine these data with patient survey, census, and electoral data to show that 1) the program is unable to fully offset the costs of care-seeking, which results in disparities in hospital utilization because some households are willing to allocate more resources to male than female health; 2) lowering costs does not reduce disparities, because males benefit as much as females do; and 3) long-term exposure to village-level female leaders reduces the gender gap in utilization, but effects are modest and limited to girls and young women. In the presence of gender bias, increasing access to and subsidizing social services may increase levels of female utilization but fail to address gender inequalities without actions that specifically target females.