WP 589- The Falling Price of Cement in Africa
The average price of cement in Africa was the highest of any continent in 2011, but by 2017 had fallen by 35 percent, the largest decline in the world, according
to data from the International Comparison Program. An empirical industry equilibrium model distinguishes between the fundamental drivers of differences in prices across countries: demand elasticity, costs, conduct, and entry. The model reveals that in 2011 average prices in Africa included the highest mark-ups and marginal costs in the world. The price decline in 2017 is accounted for by a rapid increase in the number of firms producing cement, which lowered mark-ups, and a declining marginal cost of production. Improvements in the quality of infrastructure or institutions could explain declining marginal costs, as could new technology installed by new entrants. Estimated fixed costs of entry are positive, though not systematically higher in Africa compared to other continents, suggesting equally free entry in the presence of a minimum efficient scale that does not vary across continents. These results suggest that the small size of national markets rather than regulation of entry explains why the price of cement was so high in Africa. As the market has grown, prices have fallen.