WP 588- Transhumant Pastoralism, Climate Change, and Conflict in Africa
We consider the effects of climate change on seasonally migrant populations that herd livestock — i.e., transhumant pastoralists — in Africa. Traditionally, transhumant pastoralists benefit from a cooperative relationship with sedentary agriculturalists whereby arable land is used for crop farming in the wet season and animal grazing in the dry season. Droughts can disrupt this arrangement by inducing pastoral groups to migrate to agricultural lands before the harvest, causing conflict to emerge. We examine this hypothesis by combining ethnographic information on the traditional locations of transhumant pastoralists and sedentary agriculturalists with high-resolution data on the location and timing of rainfall and violent conflict events in Africa from 1989-2018. We show that droughts in the territory of transhumant pastoralists lead to conflict in neighboring areas. Consistent with the hypothesis, these conflict events are concentrated in agricultural areas; they occur during the wet season and not the dry season; and they are due to rainfall’s impact on plant biomass growth. This mechanism explains a sizable proportion of conflict events in Africa, particularly civil conflicts and religious-extremist attacks. We find that the effects are muted in the presence of irrigation aid projects, but not in the presence of other forms of foreign aid. The effects approach zero as pastoral groups share more political power.