WP 583- Estimating Production Functions in Differentiated-Product Industries with Quantity Information and External Instruments
This paper develops a new method for estimating production-function parameters that can be applied in differentiated-product industries with endogenous quality and variety choice. We take advantage of data on physical quantities of outputs and inputs from the Colombian manufacturing survey, focusing on producers of rubber and plastic products. Assuming constant elasticities of substitution of outputs and inputs within firms, we aggregate from the firm-product to the firm level and show how quality and variety choices may bias standard estimators. Using real exchange rates and variation in the “bite” of the national minimum wage, we construct external instruments for materials and labor choices. We implement a simple two-step instrumental-variables method, first estimating a difference equation to recover the materials and labor coefficients and then estimating a levels equation to recover the capital coefficient. Under the assumption that the instruments are uncorrelated with firms’ quality and variety choices, this method yields consistent estimates, free of the quality and variety biases we have identified. Our point estimates differ from those of existing methods and changes in our preferred productivity estimator perform relatively well in predicting future export growth.