WP 582- Female Genital Cutting and the Slave Trade
We investigate the historical origins of female genital cutting (FGC), a harmful practice widespread across Africa. We test the hypothesis –substantiated by historical sources– that FGC was connected to the Red Sea slave trade route, where women were sold as concubines in the Middle East and infibulation was used to ensure chastity. We hypothesize that differential exposure of ethnic groups to the Red Sea route determined differential adoption of the practice. Combining individual level data from 28 African countries with novel historical data on slaves’ shipments by country, ethnic group and trade routes from 1400 to 1900. We find that women belonging to ethnic groups whose ancestors were exposed to the Red Sea route are more likely to be infibulated or circumcised today and are more in favor of continuing the practice. The estimated effects are very similar when slave exports are instrumented by distance to the North-Eastern African coast. Finally, the effect is smaller for ethnic groups that historically freely permitted premarital sex — a proxy for low demand for chastity.