WP 581- Industrial Clusters, Networks and Resilience to the Covid-19 Shock in China
This paper examines resilience of Chinese firms to the Covid-19 shock, and how it varied with a cluster index (measuring spatial agglomeration of firms in related industries) at the county level. Two data sources are used: entry flows of newly registered firms in the entire country, and an entrepreneur survey regarding operation of existing firms. Both datasets show greater resilience in counties with a higher cluster index, after controlling for industry dummies and local infection rates, besides county and time dummies in the entry data. We then explore possible explanations for these findings. Reliance of clusters on informal entrepreneur hometown networks provides part of the explanation. Closer proximity to suppliers and customers, and higher on-line sales help explain some of the better performance of incumbents in clusters.