WP 580- Start What You Finish! Ex Ante Risk and Schooling Investments in the Presence of Dynamic Complementarities
We study the relationship between risk and schooling investment in a low income setting, with a particular focus on possible ex ante effects. We first present a model that shows that such effects can arise if the human capital production function exhibits dynamic complementarity and parental preferences for human capital are not too concave. We then estimate the key parameters of the model using multiple rounds of panel data from rural India that contain, in each round, three seasons of time allocation for each sampled child. These estimates suggest an elasticity of schooling investments with respect to risk of -0.09 in this context. We then use cross-round differences in village-level irrigation interacted with rainfall variability to estimate the relationship between income risk and school time. Using this variation, we find an estimated elasticity of study time with respect to risk between -0.05 and -0.04. Finally, we simulate the effects of an implicit social insurance program, modeled after the National Rural Employment Guarantee Scheme (NREGS). Our results suggest that the risk-reducing effect of the NREGS may offset adverse effects on child education that were evident during the NREGS phase-in due to rising wages.