WP 575- Understanding Cultural Persistence and Change
We examine a determinant of cultural persistence that has emerged from a class of models in evolutionary anthropology: the similarity of the environment across generations. Within these models, when the environment is more similar across generations, the traits that have evolved up to the previous generation are more likely to be optimal for the current generation. In equilibrium, a greater value is placed on tradition and there is greater cultural persistence. We test this hypothesis by measuring the variability of different climatic measures across 20-year generations from 500–1900. Employing a variety of tests that use a range of samples and empirical strategies, we find that populations with ancestors who lived in environments with more cross-generational instability place less importance in maintaining tradition today and exhibit less cultural persistence.