WP 574- Optimal Subsidies for Prevention of Infectious Disease
Most economists would agree that the positive externalities caused by prevention of infectious disease create a prima facie case for subsidies. However, little is known about the appropriate magnitude of these subsidies, or about whether the level of such subsidies should vary across diseases. We integrate a standard epidemiological model with an economic model of consumer and producer behavior to address these questions. Across a continuum of market structures, we find that the equilibrium steady-state marginal externality is non-monotonic in disease transmissibility, peaking when the disease is just transmissible enough to survive in steady-state. This pattern implies that marginal externalities—and, as we show, optimal subsidies—are higher for serious but rare diseases relative to diseases with lower individual burden but higher disease prevalence. Crude calibrations suggest that optimal subsidies for technologies such as vaccines, condoms, and mosquito nets, which prevent infectious diseases, may be very large relative to current levels.