WP 573- Is Fish Brain Food or Brain Poison? Sea Surface Temperature, Methyl-mercury and Child Cognitive Development
We exploit variation in the composition of local fish catches around the time of birth using large- scale administrative and census data on adult cognitive test scores, schooling attainment, and occupation among coastal populations in Colombia to estimate the distinct causal effects of methylmercury (MeHg) and DHA, elements contained in fish, on cognitive development. Using an IV strategy based on an equilibrium model of fish supply that exploits time-series variation in oceanic SST anomalies on both coasts of Colombia from 1950 to 2014 as instruments, we find that net of cohort and municipality fixed effects increases in high-MeHg fish catches around a cohort’s birth, net of rainfall variation in the same interval, negatively affect the cohort’s verbal and math test scores upon exiting high school and their likelihood of continuing their schooling, while increasing the likelihood the cohort is disproportionally represented in manual-labor occupations. In contrast, for given levels of early-life high-MeHg fish supply, increases in early-life catch sizes of low-MeHg fish raise verbal and math test scores and educational attainment and lower the likelihood of being a manual laborer.