WP 572- How Political Insiders Lose Out When International Aid Underperforms: Evidence from a Participatory Development Experiment in Ghana
Participatory development is designed to mitigate problems of political bias in pre-existing local government but also interacts with it in complex ways. Using a five-year randomized controlled study in 97 clusters of villages (194 villages) in Ghana, we analyze the effects of a major participatory development program on participation in, leadership of and investment by pre-existing political institutions, and on households’ overall socioeconomic well-being. Applying theoretical insights on political participation and redistributive politics, we consider the possibility of both cross-institutional mobilization and displacement, and heterogeneous effects by partisanship. We find the government and its political supporters acted with high expectations for the participatory approach: treatment led to increased participation in local governance and reallocation of resources. But the results did not meet expectations, resulting in a worsening of socioeconomic wellbeing in treatment versus control villages for government supporters. This demonstrates international aid’s complex distributional consequences.