WP 571- Microentrepreneurship in Developing Countries
This article reviews the recent literature in economics on small-scale entrepreneur- ship (“microentrepreneurship”) in low-income countries. Major themes in the literature include the determinants and consequences of joining the formal sector, the impacts of access to credit and other financial services, the impacts of business training, barriers to hiring, and the distinction between self-employment by necessity and self-employment as a calling. The article devotes special attention to unique issues that arise with female entrepreneurship.