WP 569- On the Long Term Effects of the 1918 U.S. Influenza Pandemic
Using the 1918 Spanish influenza pandemic, Almond (2006) concludes that in utero exposure to maternal health insults has a large, negative impact on socio-economic status that reaches well into adulthood. A key assumption underlying this research is that birth cohorts exposed in utero to the influenza are statistically exchangeable with surrounding birth cohorts. The validity of that assumption is investigated using data from the 1920 and 1930 U.S. Censuses. We document that the exposed cohorts were born to families of lower socio-economic status relative to those who were not exposed. For example, fathers of the 1919 birth cohort were less likely to be literate, worked in lower-earning occupations, had lower socioeconomic status, were older, less likely to be white, had higher fertility and were less likely to be WWI veterans than the fathers of surrounding birth cohorts. Furthermore, after controlling for background characteristics, there is little evidence that individuals born in 1919 have worse socio-economic outcomes in adulthood relative to surrounding birth cohorts.