WP 568- Credit Rationing and Pass-Through in Supply Chains: Theory and Evidence from Bangladesh
We extend standard models of price pass-through in an imperfectly competitive supply chain to incorporate rationing of trade credit. Credit rationing reverses predictions concerning effects of raw material import prices on pass-through to wholesale prices, and effects of regulations of intermediaries. To test these we study the effects of a policy in Bangladesh’s edible oils supply chain during 2011-12 banning a layer of financing intermediaries. Evidence from a difference-in-difference estimation rejects the standard model. We find that the regulatory effort to reduce market power of financing intermediaries ended up raising consumer prices by restricting access to credit of downstream traders.