WP 567- Voting for Quality? The Impact of School Performance Information on Electoral Outcomes
We use data from polling stations and public primary schools to estimate the electoral effects of making school quality information available to voters. We exploit the introduction of a school-level accountability system in Brazil that provides, for the first time, information about school quality and exploit variation in schools affected by the policy. We find that incumbent vote shares are 2.14 percentage points higher for schools in the top 20% of our sample’s distribution of school quality and 1.54 percentage points lower for schools in the bottom 20% of the distribution. These effects are mostly driven by the unpredicted component of school quality and are larger for more educated voters.