WP 565- Housing Prices, Inter-generational Co-residence, and “Excess” Savings by the Young: Evidence using Chinese Data
In many countries of the world the co-residence of young adults aged 25-34 with their parents is not uncommon and in some countries the savings rates of these age groups exceed those of the middle-aged contrary to the standard model of life-cycle savings. In this paper we examine the role of housing prices in affecting the living arrangements of adult family members and their individual savings rates by age. Using unique data from China that enable the re-construction of whole families and identify individual savings regardless of who within the family co-resides in the same household, and exploiting the Chinese government rules determining the supply of land for residential housing, we find that increases in housing prices significantly increase inter-generational co-residence and elevate the savings rates of the young relative to the middle-aged, conditional on income, in part due to the subsidies to the young from sharing housing with parents. Based on our estimates of the effects of housing prices on co-residence and the effects of co-residence on individual savings, we find that the savings rates of the young in China would be 21% lower if housing prices were at the same ratio to disposable incomes as that observed in the United States.