WP 562- The Entertaining Way to Behavioral Change: Fighting HIV with MTV
We test the effectiveness of an entertainment education TV series, MTV Shuga, aimed at providing information and changing attitudes and behaviors related to HIV/AIDS. Using a simple model we show that “edutainment” can work through an ‘individual’ or a ‘social’ channel. We conducted a randomized controlled trial in urban Nigeria where young viewers were exposed to MTV Shuga or to a placebo TV series. Among those exposed to MTV Shuga, we created additional variation in the ‘social messages’ they received and in the people with whom they watched the show. We find significant improvements in knowledge and attitudes towards HIV and risky sexual behavior. Treated subjects are twice as likely to get tested for HIV eight months after the intervention. We also find reductions in STDs among women. These effects are stronger for viewers who report being more involved with the narrative, consistent with the psychological underpinnings of “edutainment”. Our experimental manipulations of the social norm component did not produce significantly different results from the main treatment. The ‘individual’ effect of edutainment thus seems to have prevailed in the context of our study.