WP 561- Assessing the Benefits of Long‐Run Weather Forecasting for the Rural Poor: Farmer Investments and Worker Migration in a Dynamic Equilibrium Model
The livelihoods of the majority of the world’s poor depend on agriculture. They face substantial risk from fluctuations in weather conditions. Better risk, credit and savings markets can improve productivity and welfare in rural areas but entail high administrative costs. We consider a classic public good with benefits that theoretically exceed those of perfect insurance contracts ‐ improving the skill of long‐run weather forecasts. We use an equilibrium model of agricultural production and labor migration, and a variety of Indian panel datasets to assess quantitatively the effects of improvements in seasonal forecasts of monsoon weather. We find that in areas where the forecast is accurate (has “skill”) that investment, migration and rural wages respond to forecasts. We calculate that if such skill were pervasive across India, the total value of an accurate forecast for farmers and wage workers is in the tens of billions of rupees.